requirements.txt
astroid==2.3.3
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-ckeditor==5.8.0
django-embed-video==1.3
django-js-asset==1.2.2
gunicorn==20.0.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2-binary==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
